nixos/oci-containers: Enable podman systemd integration
Podman can [auto-update containers and restart their systemd units][1]
if told what systemd unit manages the container. Podman's unit file
generation commands will set the required variable, but when using
`virtualisation.oci-containers` that doesn't happen, of course.

This commit adds that variable if the container is a podman container,
so that the auto-update feature can be used.

[1]: https://docs.podman.io/en/latest/markdown/podman-auto-update.1.html?highlight=PODMAN_SYSTEMD_UNIT#description