minecraft: Mark as broken
See #179323 for context, but in short, this launcher can no longer
launch Minecraft versions 1.19 and up, since it has switched to using
natively compiled shared libraries as part of lwjgl.

The test misses this since it doesn't actually try to launch a game.

I've not managed to find a quick workaround, so this should probably
be marked broken for now. PrismLauncher continues to work as an
alternative.